requirements.txt
faker
mysql-connector-python
transformers
accelerate